requirements.txt
functions-framework>=3.0,<4.0
Flask>=2.0.0,<3.0.0
requests
google-cloud-storage